# src/infrastructure/swagger/index.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Invernadero del rey de las flores",
    "description": "El rey de las flores tiene sus fabricas dentro de la tierra, cada obrero hace una flor que en primavera creera...",
    "version": "0.1.0"
  },
  "tags": [
    {
      "name": "flor",
      "description": ""
    }
  ],
  "paths": {
    "/flor": {
      "get": {
        "tags": [
          "flor"
        ],
        "summary": "Recupera el listado de flores disponibles y en buen estado",
        "operationId": "getFlores",
        "responses": {
          "400": {
            "description": "Bad request",
            "content": {}
          },
          "404": {
            "description": "Flor not found",
            "content": {}
          },
          "405": {
            "description": "Validation exception",
            "content": {}
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
  "/flor/{florId}": {
    "get": {
      "tags": [
        "flor"
      ],
      "summary": "Obtiene una flor especifica del invernadero",
      "operationId": "getFloresById",
      "parameters": [
        {
          "name": "florId",
          "in": "path",
          "description": "Identificador de la flor",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "successful operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/flor"
              }
            }
          }
        },
        "400": {
          "description": "Invalid flor supplied",
          "content": {}
        },
        "404": {
          "description": "flor not found",
          "content": {}
        }
      }
    }
  }
  },
  "components": {
    "schemas": {
      "flor": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "example": 357
          },
          "descripcion": {
            "type": "string",
            "example": "Peso"
          },
          "sombolo": {
            "type": "string",
            "example": "$"
          },
          "decimales": {
            "type": "integer",
            "example": 0
          }
        }
      }
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}